Separate JournalOutput, also to avoid cycles
