+++ src/CMakeLists.txt
@@ -154,6 +154,8 @@ list( APPEND SOURCES
       ImageManipulation.h
       Journal.cpp
       Journal.h
+      JournalOutput.cpp
+      JournalOutput.h
       JournalRegistry.cpp
       JournalRegistry.h
       KeyboardCapture.cpp
