+++ src/Journal.h
@@ -11,7 +11,6 @@
 #ifndef __AUDACITY_JOURNAL__
 #define __AUDACITY_JOURNAL__
 
-#include <initializer_list>
 #include "Identifier.h"
 class wxArrayString;
 class wxArrayStringEx;
@@ -31,10 +30,6 @@ namespace Journal
    //\brief Change the enablement of recording and store in preferences
    //\return whether successful
    bool SetRecordEnabled(bool value);
-
-   //\brief Whether actually recording.
-   // IsRecording() && IsReplaying() is possible
-   bool IsRecording();
 
    //\brief Whether actually replaying.
    // IsRecording() && IsReplaying() is possible
@@ -60,16 +55,6 @@ namespace Journal
    // conditions are encountered.
    // Returns true if any command was dispatched
    bool Dispatch();
-
-   //\brief write the strings to the output journal, if recording
-   // None of them may contain newlines
-   void Output( const wxString &string );
-   void Output( const wxArrayString &strings );
-   void Output( std::initializer_list< const wxString > strings );
-
-   //\brief if recording, emit a comment in the output journal that will have
-   // no effect on playback
-   void Comment( const wxString &string );
 
    //\brief If recording, output the strings; if playing back, require
    // identical strings.  None of them may contain newlines
